requirements.txt
json==2.0.9
requests==2.31.0
paramiko==2.11.0
python-dotenv==1.0.0
pandas==2.0.3